[tps] 투표 관련자료 등록시 삭제 할 일련번호 있는지 체크하는 로직 추가

+++ moka-core/moka-core-tps/src/main/java/jmnet/moka/core/tps/mvc/poll/service/TrendpollServiceImpl.java
@@ -88,8 +88,9 @@ public class TrendpollServiceImpl implements TrendpollService {
                     .stream()
                     .map(value -> value.getItemSeq())
                     .collect(Collectors.toCollection(ArrayList::new));
-
-            trendpollRepository.deleteItemByPollSeq(trendpoll.getPollSeq(), exceptSeqs);
+            if (exceptSeqs != null && exceptSeqs.size() > 0) {
+                trendpollRepository.deleteItemByPollSeq(trendpoll.getPollSeq(), exceptSeqs);
+            }
         }
         if (trendpoll.getPollRelateContents() != null && trendpoll
                 .getPollRelateContents()
@@ -99,7 +100,9 @@ public class TrendpollServiceImpl implements TrendpollService {
                     .stream()
                     .map(value -> value.getSeqNo())
                     .collect(Collectors.toCollection(ArrayList::new));
-            trendpollRepository.deleteContentsByPollSeq(trendpoll.getPollSeq(), exceptSeqs);
+            if (exceptSeqs != null && exceptSeqs.size() > 0) {
+                trendpollRepository.deleteContentsByPollSeq(trendpoll.getPollSeq(), exceptSeqs);
+            }
         }
 
         return saveTrendpollDetail(trendpoll);
